backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
redis>=5.0.4
celery>=5.4.0
anthropic>=0.26.0
python-multipart>=0.0.9
email-validator>=2.1.1
httpx>=0.27.0
pytest>=8.2.1
pytest-asyncio>=0.23.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
